Refactor to proper onion architecture with infrastructure layer separation
・Domain層をインターフェース定義のみに修正（外部依存を除去）
・Infrastructure層にapi_clients.pyを新規作成し外部API実装を分離
・WikipediaRepository → 抽象インターフェース
・WikipediaApiClient → HTTP通信の具体実装
・テスタビリティとモック化の大幅向上
・README.mdとCLAUDE.mdを新アーキテクチャに対応更新
・正しいオニオンアーキテクチャの責務分離を実現

Architecture changes:
- Domain: Abstract interfaces only (no external dependencies)
- Infrastructure: Concrete implementations (HTTP, file system)
- Usecase: Business logic composition
- UI: User interface
